requirements.txt
watchdog>=3.0.0
pystray>=0.19.5
Pillow>=10.0.0

